+++ src/main/java/com/example/jammoney/news/dto/NewsDto.java
@@ -1,4 +1,4 @@
-package com.example.jammoney.News.dto;
+package com.example.jammoney.news.dto;
 
 import java.time.LocalDateTime;
 import java.util.List;
