+++ src/main/java/com/example/jammoney/news/entity/NewsQuiz.java
@@ -1,0 +1,30 @@
+package com.example.jammoney.news.entity;
+
+import jakarta.persistence.*;
+import lombok.*;
+
+import java.util.List;
+
+@Entity
+@Table(name = "news_quizzes")
+@Getter
+@Setter
+@NoArgsConstructor
+@AllArgsConstructor
+@Builder
+public class NewsQuiz {
+
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
+    private Long id;
+
+    private String question;
+    private String correctAnswer;
+
+    @ElementCollection
+    private List<String> choices;
+
+    @ManyToOne
+    @JoinColumn(name = "news_id")
+    private News news;
+}
