+++ src/main/java/com/example/jammoney/theme/dto/ThemeDto.java
@@ -1,4 +1,4 @@
-package com.example.jammoney.ThemeLearning.dto;
+package com.example.jammoney.theme.dto;
 
 import lombok.Getter;
 import lombok.Setter;
