+++ src/main/java/com/example/jammoney/theme/dto/TopicDetailDto.java
@@ -1,12 +1,13 @@
-package com.example.jammoney.ThemeLearning.dto;
+package com.example.jammoney.theme.dto;
 
 import lombok.Getter;
 import lombok.Setter;
 
 @Getter
 @Setter
-public class TopicListDto {
+public class TopicDetailDto {
     private Long topicId;
     private String title;
+    private String description;
 }
