+++ src/main/java/com/example/jammoney/theme/service/LearningTopicService.java
@@ -1,12 +1,11 @@
-package com.example.jammoney.ThemeLearning.service;
+package com.example.jammoney.theme.service;
 
-import com.example.jammoney.ThemeLearning.dto.ThemeDto;
-import com.example.jammoney.ThemeLearning.dto.TopicDetailDto;
-import com.example.jammoney.ThemeLearning.dto.TopicListDto;
-import com.example.jammoney.ThemeLearning.models.LearningTopic;
-import com.example.jammoney.ThemeLearning.models.Theme;
-import com.example.jammoney.ThemeLearning.repository.LearningTopicRepository;
-import com.example.jammoney.ThemeLearning.repository.ThemeRepository;
+import com.example.jammoney.theme.dto.TopicDetailDto;
+import com.example.jammoney.theme.dto.TopicListDto;
+import com.example.jammoney.theme.entity.LearningTopic;
+import com.example.jammoney.theme.entity.Theme;
+import com.example.jammoney.theme.repository.LearningTopicRepository;
+import com.example.jammoney.theme.repository.ThemeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
