+++ src/main/java/com/example/jammoney/theme/service/ThemeService.java
@@ -1,7 +1,7 @@
-package com.example.jammoney.ThemeLearning.service;
+package com.example.jammoney.theme.service;
 
-import com.example.jammoney.ThemeLearning.dto.ThemeDto;
-import com.example.jammoney.ThemeLearning.repository.ThemeRepository;
+import com.example.jammoney.theme.dto.ThemeDto;
+import com.example.jammoney.theme.repository.ThemeRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
